requirements.txt
# Dependencias principales
PySide6
SQLAlchemy
mysql-connector-python
python-dotenv

# Dependencias opcionales (funcionalidades extra)
# pandas + openpyxl -> exportar a .xlsx
pandas
openpyxl
# reportlab -> exportar a PDF
reportlab

# Migraciones de esquema
alembic

# Si quieres desplegar la aplicacion como ejecutable
pyinstaller 

#Si quieres configurar las opciones del tipo modifica en la carpeta config